some stories refactor
