+++ packages/react-components/src/components/Tooltip/Tooltip.stories.css
@@ -1,12 +1,31 @@
 .tooltip-icon-wrap {
-  align-items: center;
-  border-radius: 4px;
   display: flex;
   flex-direction: row;
-  height: 32px;
+  align-items: center;
   justify-content: center;
+  border-radius: 4px;
   width: 32px;
+  height: 32px;
 }
+
 .tooltip-icon-wrap:hover {
   background-color: #e5e5e5;
+}
+
+.tooltip-preview-container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 800px;
+}
+
+.tooltip-preview-reports {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border: 1px dashed var(--border-default);
+  border-radius: 4px;
+  background-color: var(--surface-basic-default);
+  width: 100%;
+  height: 100px;
 }
