+++ packages/react-components/src/components/Tooltip/components/Interactive.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
-import { Icon } from '../Icon';
+import { Icon } from '../../Icon';
 import { Close } from '@livechat/design-system-icons/react/material';
-import { Button, ButtonKind } from '../Button';
-import { getIconType } from './helpers';
-import styles from './Tooltip.module.scss';
+import { Button, ButtonKind } from '../../Button';
+import { getIconType } from '../helpers';
+import styles from '../Tooltip.module.scss';
 
 const baseClass = 'tooltip';
