+++ packages/react-components/src/components/Tooltip/components/UserGuideStep.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import cx from 'clsx';
-import { Icon } from '../Icon';
+import { Icon } from '../../Icon';
 import { Close } from '@livechat/design-system-icons/react/material';
-import { Button } from '../Button';
-import { getIconType } from './helpers';
-import styles from './Tooltip.module.scss';
+import { Button } from '../../Button';
+import { getIconType } from '../helpers';
+import styles from '../Tooltip.module.scss';
 
 const baseClass = 'tooltip';
