+++ packages/react-components/src/components/Tooltip/components/index.ts
@@ -1,0 +1,5 @@
+export { Simple } from './Simple';
+export { Info } from './Info';
+export { Interactive } from './Interactive';
+export { UserGuide } from './UserGuide';
+export { Reports } from './Reports';
