+++ packages/react-components/src/components/Tooltip/index.ts
@@ -1,5 +1,2 @@
 export { Tooltip } from './Tooltip';
-export { Simple } from './Simple';
-export { Info } from './Info';
-export { Interactive } from './Interactive';
-export { UserGuide } from './UserGuide';
+export { Simple, Info, Interactive, UserGuide, Reports } from './components';
